refactor: move client functions to hooks.ts
