+++ frontend/src/components/ActorEntryForm.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useRef, useState } from "react";
-import { ActorBase, useOptions, toSelectOptions } from "@/app/(system)/common";
-import { useObjectState } from "@/app/(system)/hooks";
+import { ActorBase, toSelectOptions } from "@/app/(system)/common";
+import { useObjectState, useOptions } from "@/app/(system)/hooks";
 import { useTranslation } from "@/app/(system)/internationalization";
 import {
   FieldErrorContext,
