+++ frontend/src/components/LiceneseEntryForm.tsx
@@ -2,10 +2,9 @@ import {
   ActorBase,
   LicenseInstance,
   Options,
-  useOptions,
   toSelectOptions,
 } from "@/app/(system)/common";
-import { useFilter, useObjectState } from "@/app/(system)/hooks";
+import { useFilter, useObjectState, useOptions } from "@/app/(system)/hooks";
 import { useTranslation } from "@/app/(system)/internationalization";
 import {
   FieldErrorContext,
